movielens_api/requirements.txt
Django>=4.0,<5.0
djangorestframework
drf-yasg
pytest
pytest-django
